refactor(section): Section에서 상향,하향 종점 여부를 관리하도록 수정
